THIS IS BROKEN... Late night idea, screwed it up, but leaving it at...
I know what I am going to do. I'm just going to create a string buffer for
the data and convert it to a Char array just before I need to transmit it.
So leaving this mess of code here and this commit message so I know what
my plan of action for this code is for tomorrow.

But something I watch is starting in 15 mins so going to leave this for
the night and settle in.