Impletemented a simple CRC check
The Pi sends out data to the RF module near the front door, The module
extacts the last two bytes from the payload and uses it for a CRC Check.
The rest of the payload is checked against this CRC and its success or
failure is reported back to the pi, the pi will either empty the buffer
(on a success) or retransmit the buffer upto 3 times before giving up
(don't want to get stuck in a endless loop). This is all done on the
micros.