I need to settle on a format for the LCD with some control chars
Do I want the screen to line wrap or just stop printing when at a end of a
line?

Do I want to skip spaces at the start of a line or not (in line wrap it
would be useful but it won't allow me to "center" text with spaces
(Note to self: I also want a center text control char)

Do I want to ignore new lines at the start of a line?

This is going to need some pen and paper :-)